fix some warnings and add debug info to unit_test build type

+++ usr/RL/src/ActiveOnlineSearch/DynamicTightRope.h
@@ -21,7 +21,7 @@ public:
     virtual QString action_name(const action_t &) const override;
     virtual QString state_name(const state_t &) const override;
     bool has_terminal_state() const override {return true;}
-    bool is_terminal_state(state_t s) const override {return s==states.size()-1;}
+    bool is_terminal_state(state_t s) const override {return s==(int)states.size()-1;}
     std::tuple<int,int> get_position_and_velocity(const state_t & state) const;
     DynamicTightRope(int n = 15);
     virtual ~DynamicTightRope() = default;
